% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/engine.R
\name{active_project}
\alias{active_project}
\title{The current active Syberia project.}
\usage{
active_project()
}
\value{
A \code{syberia_engine} object or \code{NULL}.
}
\description{
The current active Syberia project.
}

