% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{any_is_substring_of}
\alias{any_is_substring_of}
\title{Whether or not any substring of a string is any of a set of strings.}
\usage{
any_is_substring_of(string, set_of_strings)
}
\arguments{
\item{string}{character.}

\item{set_of_strings}{character.}
}
\value{
logical
}
\description{
Whether or not any substring of a string is any of a set of strings.
}
\examples{
stopifnot(syberia:::any_is_substring_of('test', c('blah', 'te', 'woo'))) # TRUE
stopifnot(!syberia:::any_is_substring_of('test', c('blah', 'woo'))) # FALSE
}

