% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{auto_run}
\alias{auto_run}
\title{Auto run a stage in the last syberia active runner.}
\usage{
auto_run(x, ...)
}
\arguments{
\item{x}{expression. The last executed R expression. If of the form
\code{a / b}, it will translate to \code{active_runner()$run(, "a/b")}.}

\item{...}{additional arguments the task callback usually receives.}
}
\description{
This is used in conjunction with \code{base::addTaskCallback} when running
a model from the interactive console.
}
\examples{
# From console, type:
# > 2/5
# and the fifth substage of the second stage of the last active runner
# will be executed.
}
\seealso{
\code{\link{addTaskCallback}}
}

