% Generated by roxygen2 (5.0.1.9000): do not edit by hand
% Please edit documentation in R/engine.R
\name{build_engine}
\alias{build_engine}
\title{Build a syberia engine.}
\usage{
build_engine(buildable)
}
\arguments{
\item{buildable}{character. A file path from which to build the engine.}
}
\description{
Build a syberia engine.
}

