\name{build_export_stagerunner}
\alias{build_export_stagerunner}
\title{Build a stagerunner for exporting data with backup sources.}
\usage{
build_export_stagerunner(modelenv, export_options, meta_options = list())
}
\arguments{
  \item{modelenv}{an environment. The current modeling
  environment.}

  \item{export_options}{list.}

  \item{meta_options}{list.}
}
\description{
Build a stagerunner for exporting data with backup sources.
}

