\name{build_import_stagerunner}
\alias{build_import_stagerunner}
\title{Build a stagerunner for importing data with backup sources.}
\usage{
build_import_stagerunner(modelenv, import_options)
}
\arguments{
  \item{modelenv}{an environment. The current modeling
  environment.}

  \item{import_options}{a list. Nested list, one adapter
  per list entry.}
}
\description{
Build a stagerunner for importing data with backup sources.
}

