% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/director.R
\name{default_tests_environment_preprocessor}
\alias{default_tests_environment_preprocessor}
\title{The default preprocessor for syberia test environments.}
\usage{
default_tests_environment_preprocessor(director, source_args, source)
}
\arguments{
\item{director}{director}

\item{source_args}{list}

\item{source}{function}
}
\description{
The default preprocessor for syberia test environments.
}

