% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ensure_no_global_variable_pollution}
\alias{ensure_no_global_variable_pollution}
\title{Ensure no global variables are polluted during an expression.}
\usage{
ensure_no_global_variable_pollution(expr, desc)
}
\arguments{
\item{expr}{expression. The R expression to evaluate}

\item{desc}{character. A string to add to "you modified global
variables while [\code{desc} goes here]".}
}
\description{
If any global variables are removed or created, it will
give a descriptive error.
}

