% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/engine_tests.R
\name{ensure_resources_are_tested}
\alias{ensure_resources_are_tested}
\title{Check that all mandatory tested resources have tests.}
\usage{
ensure_resources_are_tested(engine, tests, optional)
}
\arguments{
\item{engine}{syberia_engine. The engine to check.}

\item{tests}{character. The tests to check. Must be a list with keys
\code{"active"} and \code{"ignored"}.}

\item{optional}{character. A character vector of optional tests.}
}
\value{
Nothing, but error if not all resources have tests.
}
\description{
Check that all mandatory tested resources have tests.
}

