% Generated by roxygen2 (4.0.0): do not edit by hand
\name{export_adapter}
\alias{export_adapter}
\title{Fetch an export adapter.}
\usage{
export_adapter(adapter = "file")
}
\arguments{
\item{adapter}{character. Only supported so far are 's3' and 'file'.
The default is 'file'.}

\item{opts}{list. The options that get passed to the export adapter.}
}
\description{
Fetch an export adapter.
}

