% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fetch_model_container}
\alias{fetch_model_container}
\title{Fetch a tundra model container.}
\usage{
fetch_model_container(type)
}
\arguments{
\item{type}{character. The model keyword. This function will attempt
to fetch the associated container construction function (with
parameters \code{munge_procedure}, \code{default_args}, and
\code{internal} from either (1) the tundra package, or (2)
your syberia project's \code{lib/classifiers} directory (see
description).}
}
\value{
A container construction function (with
  parameters \code{munge_procedure}, \code{default_args}, and
  \code{internal} which takes these and return a \code{tundra_container}
  object with a \code{train} and \code{predict} method.
}
\description{
The container is fetched either from the tundra package or from
the \code{lib/classifiers} directory in the syberia project.
}
\details{
If there is no tundra model associated to the keyword (like 'gbm'
or 'regularization'), you must place a function called \code{train}
and a function called \code{predict} in the an R file with the same
name as the keyword for your classification. For example,
if you are implementing least-angle regression, you could define
\code{lib/classifiers/lar.R} and from your syberia model use
\code{'lar'} for your model keyword.
}

