% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{find_test_hook}
\alias{find_test_hook}
\title{Fetch the test setup or teardown hook, if any exists.}
\usage{
find_test_hook(engine, type = "setup", config)
}
\arguments{
\item{engine}{syberia_engine. The director for the syberia project.}

\item{type}{character. Must be \code{'setup'} or \code{'teardown'}, the former
being the default.}

\item{config}{character. The resource used to fetch configuration.}
}
\value{
a stageRunner that will run the relevant setup or teardown hook(s).
}
\description{
The resource \code{config/environments/test} should contain a local variable
\code{setup} or \code{teardown} that has a function or list of functions to
be incorporated into a stageRunner that will run the actual test setup
or teardown.
}
\details{
The seed environment for the stageRunner will contain the director object
of the relevant project in the key \code{director}.
}
\seealso{
\code{\link{test_engine}}
}

