\name{import_stage}
\alias{import_stage}
\title{Import data stage for Syberia model process.}
\usage{
import_stage(import_options)
}
\arguments{
  \item{import_options}{a list. The available import
  options. Will differ depending on the adapter. (default
  is file adapter)}
}
\description{
Import data stage for Syberia model process.
}

