% Generated by roxygen2 (5.0.1.9000): do not edit by hand
% Please edit documentation in R/utils.r
\name{infix_or}
\alias{\%||\%}
\alias{infix_or}
\title{Non-null selection operator.}
\usage{
x \%||\% y
}
\arguments{
\item{x}{ANY. An R object. If \code{NULL}, return \code{y}.}

\item{y}{ANY. An R object. If x is \code{NULL}, return \code{y}.}
}
\description{
Non-null selection operator.
}

