% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{install_github_packages}
\alias{install_github_packages}
\title{Install packages from github}
\usage{
install_github_packages(pkgs, apply_method = lapply)
}
\arguments{
\item{pkgs}{a list of packages that get passed as arguments into
devtools::install_github}

\item{apply_method}{a function. Useful for custom progress decorators. The
default is \code{lapply}.}
}
\description{
Install packages from github
}
\examples{
\dontrun{
install_github_packages(list(list('Ramd', 'robertzk'),
  list('s3mpi', 'robertzk', 'rconference')))
}
}
\seealso{
\code{\link{load_github_packages}}
}

