% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install_syberia.R
\name{install_syberia}
\alias{install_syberia}
\title{Install syberia.io distribution on your machine.}
\usage{
install_syberia()
}
\description{
Install syberia.io distribution on your machine.
}

