% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/convenience.R
\name{last_run}
\alias{last_run}
\title{The before and after environments of the last syberia run.}
\usage{
last_run()
}
\value{
a list with \code{before} and \code{after} keys giving
   what the modeling environment looked like before and after
   the last syberia run.
}
\description{
The before and after environments of the last syberia run.
}

