\name{model_stage}
\alias{model_stage}
\title{Model stage for syberia models}
\usage{
model_stage(modelenv, model_parameters)
}
\arguments{
  \item{modelenv}{an environment. The persistent modeling
  environment.}

  \item{model_parameters}{a list. Model-specific
  parameters, with the first parameter always being the
  model keyword for the tundra container (e.g., glm, gbm,
  etc.)}
}
\description{
TODO: Document this more
}

