% Generated by roxygen2 (5.0.1.9000): do not edit by hand
% Please edit documentation in R/director.R
\name{register_config}
\alias{register_config}
\title{Register configuration when bootstrapping a syberia project.}
\usage{
register_config(project)
}
\arguments{
\item{project}{director. The syberia director object to bootstrap.}
}
\description{
Register configuration when bootstrapping a syberia project.
}

