% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/director.R
\name{register_controllers}
\alias{register_controllers}
\title{Register parser for controllers.}
\usage{
register_controllers(project)
}
\arguments{
\item{project}{director. The syberia director object on which to register
the default controllers parser.}
}
\description{
The only non-trivial action is to look for a \code{preprocessors} local
variable in the \code{input} provider to the parser. The parsed value
of a controller will be a list containing a \code{preprocessor} and
a \code{parser}
}

