% Generated by roxygen2 (4.0.1): do not edit by hand
\name{register_routes}
\alias{register_routes}
\title{Register routes when bootstrapping a syberia project.}
\usage{
register_routes(project)
}
\arguments{
\item{project}{director. The syberia director object to bootstrap.}
}
\description{
Register routes when bootstrapping a syberia project.
}

