% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/director.R
\name{routes_parser}
\alias{routes_parser}
\title{A director parser for parsing a routes file.}
\usage{
routes_parser()
}
\description{
A routes file (in \code{"config/routes.r"} relative to the syberia project)
should contain a list whose names give the route prefixes and whose
values are the controller names for parsing resources that begin with
these prefixes. For example, if the file contains
}
\details{
\code{list('models' = 'models', 'lib/adapters' = 'adapters')}

then files in the directory \code{"models"} (relative to the root of the
syberia project) will be processed by the \code{"models"} controller
and the files in the directory \code{"lib/adapters"} will be
processed by the \code{adapters} controller. There should be files
\code{"adapters.R"} and \code{"models.R"} in the \code{"lib/controllers"}
directory containing only a single function. The function (the heart of
the controller) will be applied to files sourced in the respective
resource directory (\code{"models"} or \code{"lib/adapters"}).
See the function \code{bootstrap_syberia_project} to understand what
things are available in a controller function.
}
\seealso{
\code{bootstrap_syberia_project}.
}

