% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.r
\name{run_model}
\alias{run_model}
\title{Build a model using a data source from scratch.}
\usage{
run_model(key = syberiaStructure:::get_cache("last_model") \%||\%
  getOption("syberia.default_model"), ..., fresh = FALSE, verbose = TRUE,
  params = list())
}
\arguments{
\item{key}{a string or list. If the former, there must be a
file with name \code{model_stages} followed by \code{.r} so that syberia
can read the model configurations.}

\item{...}{additional arguments to pass to \code{$run(...)} on the stageRunner.
For example, \code{to = 'some/key'}.}

\item{fresh}{logical. Whether or not to use the cache. By default, \code{FALSE}.}

\item{verbose}{logical. Whether or not to display messages. The default is
\code{TRUE}.}

\item{params}{list. Inject parameters into \code{modelenv}. You will have to implement
the handling of those parameters yourself.
For example, \code{run('some_file', params=list(invalidate_cache = TRUE))}.}
}
\description{
Build a model using a data source from scratch.
}

