% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/engine.R
\name{should_exclude}
\alias{should_exclude}
\title{Whether to exclude a syberia engine from being used for resourcing.}
\usage{
should_exclude(condition, engine)
}
\arguments{
\item{condition}{logical. Some condition.}

\item{engine}{syberia_engine. Engine object.}
}
\description{
Whether to exclude a syberia engine from being used for resourcing.
}

