% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{syberia}
\alias{syberia}
\alias{syberia-package}
\title{Syberia provides an opinionated unified framework for
fast iteration on classifier development and deployment. It is
founded on convention over configuration and aims to solve the
problems of classifier-specific data preparation and
classifier-specific modeling parameters.}
\description{
Syberia provides an opinionated unified framework for
fast iteration on classifier development and deployment. It is
founded on convention over configuration and aims to solve the
problems of classifier-specific data preparation and
classifier-specific modeling parameters.
}

