% Generated by roxygen2 (4.0.0): do not edit by hand
\name{syberia_config}
\alias{syberia_config}
\title{Fetch the configuration for a Syberia file.}
\usage{
syberia_config(root = syberia_root(), exists_check = FALSE)
}
\arguments{
\item{root}{character. The root of the Syberia project. The
default is \code{syberia_root()}.}

\item{exists_check}{logical. If \code{TRUE}, it will only return
whether or not the directory contains a syberia configuration
file (either \code{syberia.config} or \code{syberia_config.r},
with the extensions case-insensensitive, .r or .R). The default
is \code{FALSE}: The whole syberia configuration will be returned
for this project as a list.}
}
\value{
the syberia configuration. All local variables set in
  the syberia configuration file will be part of a list of options
  for this project. TODO: (RK) Record list of possible options somewhere.
}
\description{
Fetch the configuration for a Syberia file.
}

