% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{syberia_new}
\alias{syberia_new}
\title{Bootstrap a new syberia project}
\usage{
syberia_new(name)
}
\arguments{
\item{name}{character. The name of the folder where the project should be created.}
}
\description{
Bootstrap a new syberia project
}

