% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{syberia_new_classifier}
\alias{syberia_new_classifier}
\title{Bootstrap a new classifier for a syberia project}
\usage{
syberia_new_classifier(name)
}
\arguments{
\item{name}{character. The name of the classifier}
}
\description{
Bootstrap a new classifier for a syberia project
}

