% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{syberia_new_model}
\alias{syberia_new_model}
\title{Bootstrap a new model for a syberia project}
\usage{
syberia_new_model(name)
}
\arguments{
\item{name}{character. The name of the model.}
}
\description{
Bootstrap a new model for a syberia project
}

