% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{syberia_new_mungebit}
\alias{syberia_new_mungebit}
\title{Bootstrap a new mungebit for a syberia project}
\usage{
syberia_new_mungebit(name)
}
\arguments{
\item{name}{character. The name of the mungebit.}
}
\description{
Bootstrap a new mungebit for a syberia project
}

