% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/director.R
\name{syberia_project}
\alias{syberia_project}
\title{Fetch a syberia project director relative to a filename.}
\usage{
syberia_project(filename = getwd())
}
\arguments{
\item{filename}{character. Some in a syberia project.}
}
\value{
the director object for the syberia project.
}
\description{
Fetch a syberia project director relative to a filename.
}
\examples{
\dontrun{
  # Pretend you have the file blah/foo.R in your syberia project ~/proj.
  syberia_projects('~/proj/blah/foo.R')
  # Now we have the director object for the ~/proj syberia project.
}
}

