% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{syberia_version}
\alias{syberia_version}
\title{Fetch the current Syberia version.}
\usage{
syberia_version()
}
\value{
The version of the syberia package as \code{\link{package_version}}
  object.
}
\description{
Fetch the current Syberia version.
}

