% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/engine_tests.R
\name{test_engine}
\alias{test_engine}
\title{Run all tests in a syberia project or engine.}
\usage{
test_engine(engine = syberia_engine(), base = "test",
  config = file.path("config", "environments", "test"),
  ignored_tests = ignored_tests_from_config(engine, base, config),
  required = TRUE)
}
\arguments{
\item{engine}{syberia_engine. The syberia engine to test.
If a \code{character}, it will be passed to \code{\link{syberia_engine}} first.}

\item{base}{character. Any subdirectory to test specifically. By default,
\code{"test"}.}

\item{config}{character. The relative path to the configuration resource,
by default \code{"config/environments/test"}.}

\item{ignored_tests}{character. The list of tests to ignore, by default
the local variable \code{ignored_tests} extracted from the configuration
resource specific by the \code{config} parameter.}

\item{required}{logical. Whether or not all tests are required to have resources,
by default \code{TRUE}. If \code{TRUE}, the \code{ignored_tests}
resources will not be required to have an accompanying test.}
}
\value{
\code{TRUE} if all tests pass or will error otherwise. Note this
   function uses \code{pblapply} from the pbapply package to
   represent progress if available.
}
\description{
The tests that will be run are all those in the \code{test} subdirectory
of the root of the syberia engine, unless otherwise specified.
}
\details{
It is possible to introduce additional behavior prior to and after tests.
This can be used to perform additional testing not covered by sourcing
all files in the "test/" directory of the syberia engine.

To provide a setup or teardown hook, simply place a function or list of
functions in a local variable \code{setup} or \code{teardown}, respectively,
in \code{config/environments/test} relative to the root of the syberia engine,
or pass the relevant \code{setup} or \code{teardown} parameters to this function.

For example, creating a file \code{config/environments/test.R} with the code
\code{setup <- function(env) cat("Running all tests.")} will print a message
before all the tests are run. The one parameter the function must take is an
environment which will contain a single key, \code{director}, pointing to the
object returned by calling \code{\link{syberia_engine}}.
}
\seealso{
\code{\link{syberia_engine}}
}

