% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test_environment_config}
\alias{test_environment_config}
\title{Get the configuration for the test environment.}
\usage{
test_environment_config(project)
}
\arguments{
\item{project}{director or character. The director for the syberia project.}
}
\description{
Get the configuration for the test environment.
}

