% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{test_hook}
\alias{test_hook}
\title{Fetch the test setup or teardown hook, if any exists.}
\usage{
test_hook(project, type = "setup")
}
\arguments{
\item{project}{director or character. The director for the syberia project.}

\item{type}{character. Must be \code{'setup'} or \code{'teardown'}, the former
being the default.}
}
\value{
a stageRunner that will run the relevant setup or teardown hook(s).
}
\description{
The resource \code{config/environments/test} should contain a local variable
\code{setup} or \code{teardown} that has a function or list of functions to
be incorporated into a stageRunner that will run the actual test setup
or teardown.
}
\details{
The seed environment for the stageRunner will contain the director object
of the relevant project in the key \code{director}.
}
\seealso{
\code{\link{test_project}}
}

