% Generated by roxygen2 (4.0.0): do not edit by hand
\name{test_project}
\alias{test_project}
\title{Run all tests in a syberia project.}
\usage{
test_project(project)
}
\arguments{
\item{project}{director or character. The director for the syberia project.
If a \code{character}, it will be passed to \code{syberia_project} first.}
}
\value{
\code{TRUE} if all tests pass or will error otherwise. Note this
   function uses \code{pblapply} from the \code{pbapply} package to
   represent progress.
}
\description{
TODO: (RK) Define custom tests for invisible resources.
}
\seealso{
\code{\link{syberia_project}}
}

