% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.R
\name{test_resource}
\alias{test_resource}
\title{Run the tests for a single resource.}
\usage{
test_resource(engine, resource, setup, teardown, reporter)
}
\arguments{
\item{engine}{syberia_engine. The engine to run the test on.}

\item{resource}{character. The resource to test.}

\item{setup}{stageRunner. A \code{\link[stagerunner]{stageRunner}} to
execute setup hooks for this test.}

\item{teardown}{stageRunner. A \code{\link[stagerunner]{stageRunner}} to
execute teardown hooks for this test.}

\item{reporter}{reporter. A testthat reporter object.}
}
\value{
The testthat result summary for this one test run.
}
\description{
Run the tests for a single resource.
}

