% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/engine_tests.R
\name{test_resources}
\alias{test_resources}
\title{Run the tests on a given set of resources.}
\usage{
test_resources(engine, tests, ...)
}
\arguments{
\item{engine}{syberia_engine. The engine to run the tests on.}

\item{tests}{character. The character vector of resources to test.}

\item{...}{Additional arguments to pass to \code{find_test_hook}.}
}
\value{
The testthat result summary for this one test run.
}
\description{
Run the tests on a given set of resources.
}

