% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{traverse_parent_directories}
\alias{traverse_parent_directories}
\title{Perform an action repeatedly on parent directories until success or error.}
\usage{
traverse_parent_directories(filepath, fn, error)
}
\arguments{
\item{filepath}{character. The filepath to traverse along. The \code{fn}
function will be called with \code{filepath} and its parent directories
until it returns a result other than \code{NULL}.}

\item{fn}{function. A one-argument function called on \code{filepath} or
its successive parent directories until a result other than \code{NULL}
is returned, which will be the final return value.}

\item{error}{character or function. A string to error if \code{fn} returns
\code{NULL} on all parent directories, or a one-argument function to
execute (the argument received will be the initial \code{filepath})}
}
\value{
The result of \code{fn} on the first parent directory of
  \code{filepath} on which it is not \code{NULL}.
}
\description{
Given a \code{fn}, we may wish to run it on a \code{filepath}, determine
its success, and try again with the parent directory of \code{filepath},
until we obtain result that is not \code{NULL}. If this does not occur for
any parent directory, we halt with the string \code{error}.
}

