% Generated by roxygen2 (4.0.0): do not edit by hand
\name{build_export_stagerunner}
\alias{build_export_stagerunner}
\title{Build a stagerunner for exporting data with backup sources.}
\usage{
build_export_stagerunner(export_options, meta_options = list())
}
\arguments{
\item{export}{a list. Nested list, one adapter per list entry.}
}
\description{
Build a stagerunner for exporting data with backup sources.
}

