% Generated by roxygen2 (4.0.0): do not edit by hand
\name{build_import_stagerunner}
\alias{build_import_stagerunner}
\title{Build a stagerunner for importing data with backup sources.}
\usage{
build_import_stagerunner(import_options, meta_options = list())
}
\arguments{
\item{import_options}{list. Nested list, one adapter per list entry.}

\item{meta_options}{list. Any additional special arguments. Currently,
only \code{skip} is a supported value. If this key is in the list,
then its value will be used to try to load the data from a global
variable with that name.}
}
\description{
Build a stagerunner for importing data with backup sources.
}

