\name{build_model}
\alias{build_model}
\title{Build a model using a data source from scratch.}
\usage{
build_model(key)
}
\arguments{
  \item{key}{a string. There must be a file with name
  \code{key} followed by '.r' so that syberia can read the
  model configurations.}
}
\description{
Build a model using a data source from scratch.
}

