% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cache}
\alias{cache}
\title{Simple caching mechanism
http://r.789695.n4.nabble.com/what-is-the-preferred-method-to-create-a-package-local-variable-tp923040p923052.html}
\usage{
cache()
}
\description{
Simple caching mechanism
http://r.789695.n4.nabble.com/what-is-the-preferred-method-to-create-a-package-local-variable-tp923040p923052.html
}

