% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/construct_stage_runner.r
\name{construct_stage_runner}
\alias{construct_stage_runner}
\title{Return a stageRunner object that parametrizes a list of stages.}
\usage{
construct_stage_runner(stages, modelenv = new.env())
}
\arguments{
\item{stages}{a list of lists. Each sublist is the argument to its
named stage.}

\item{modelenv}{environment. The modeling environment with which to
construct the stagerunner. All operations will be performed on this
environment. The default is simply \code{new.env()}.}
}
\value{
stageRunner parametrizing the given stages
}
\description{
Each stage is first fed through a function that converts it to a stageRunner
or a function. For example, list(import = X, ...) gets converted by,
amongst other things, looking for a variable import_stage and passing
in X.
}
\details{
This kind of procedure allows one to define new stage types dynamically,
and "reach in" and execute only sub-parts of stages by using stageRunner
effectively.
}

