% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/director.R
\name{default_tests_preprocessor}
\alias{default_tests_preprocessor}
\title{The default preprocessor for syberia tests.}
\usage{
default_tests_preprocessor(resource_object, director, source_args, source)
}
\arguments{
\item{resource_object}{directorResource}

\item{director}{director}

\item{source_args}{list}

\item{source}{function}
}
\description{
The default preprocessor for syberia tests.
}

