% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/director.R
\name{default_tests_preprocessor}
\alias{default_tests_preprocessor}
\title{The default preprocessor for syberia tests.}
\usage{
default_tests_preprocessor(resource, director, source_env, source)
}
\arguments{
\item{resource}{character}

\item{director}{director}

\item{source_env}{list}

\item{source}{function}
}
\description{
The default preprocessor for syberia tests.
}

