% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{ensure_resources_are_tested}
\alias{ensure_resources_are_tested}
\title{Check that all mandatory tested resources have tests.}
\usage{
ensure_resources_are_tested(engine, tests, optional, base = "test")
}
\arguments{
\item{engine}{syberia_engine. The engine to check.}

\item{tests}{character. The tests to check. Must be a list with keys
\code{"active"} and \code{"ignored"}.}

\item{optional}{character. A character vector of optional tests.}

\item{base}{character. The directory containing tests in the project, by
default \code{"test"}.}
}
\value{
Nothing, but error if not all resources have tests.
}
\description{
Check that all mandatory tested resources have tests.
}

