% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{ensure_resources_have_tests}
\alias{ensure_resources_have_tests}
\title{Check that all mandatory tested resources have tests.}
\usage{
ensure_resources_have_tests(project, tests = project$find(base = "test/"),
  ignore = character(0), test_config)
}
\arguments{
\item{project}{director or character. The director for the syberia project.}

\item{tests}{character. The tests to check. By default, all tests in the project.}

\item{ignore}{character. A vector of tests to ignore (and not check for presence).}

\item{test_config}{environment or list. Test configuration for the project,
by default the \code{"config/environments/test"} resource.}
}
\description{
Check that all mandatory tested resources have tests.
}

