% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/tests.r
\name{ensure_resources_have_tests}
\alias{ensure_resources_have_tests}
\title{Check that all mandatory tested resources have tests.}
\usage{
ensure_resources_have_tests(project, tests = project$find(base = "test/"),
  ignore = character(0))
}
\arguments{
\item{project}{director or character. The director for the syberia project.}

\item{tests}{character. The tests to check. By default, all tests in the project.}

\item{ignore}{character. A vector of tests to ignore (and not check for presence).}
}
\description{
Check that all mandatory tested resources have tests.
}

