\name{export_stage}
\alias{export_stage}
\title{Export stage for Syberia.}
\usage{
export_stage(export_options)
}
\arguments{
  \item{export_options}{a list. The available export
  options. Will differ depending on the adapter. (default
  is file adapter)}
}
\description{
Precise behavior depends on adapter.
}

