% Generated by roxygen2 (4.0.0): do not edit by hand
\name{import_adapter}
\alias{import_adapter}
\title{Fetch an import adapter.}
\usage{
import_adapter(adapter = "file")
}
\arguments{
\item{adapter}{character. Only supported so far are 's3' and 'file'.
The default is 'file'.}

\item{opts}{list. The options that get passed to the import adapter.}
}
\description{
Fetch an import adapter.
}

