% Generated by roxygen2 (4.0.0): do not edit by hand
\name{model_stage}
\alias{model_stage}
\title{Model stage for syberia models}
\usage{
model_stage(model_parameters)
}
\arguments{
\item{model_parameters}{a list. Model-specific parameters, with the first
parameter always being the model keyword for the tundra container
(e.g., glm, gbm, etc.)}
}
\description{
TODO: Document this more
}

