% Generated by roxygen2 (4.0.1): do not edit by hand
\name{register_config}
\alias{register_config}
\title{Register configuration when bootstrapping a syberia project.}
\usage{
register_config(project)
}
\arguments{
\item{project}{director. The syberia director object to bootstrap.}
}
\description{
Register configuration when bootstrapping a syberia project.
}

