% Generated by roxygen2 (4.0.0): do not edit by hand
\name{register_tests}
\alias{register_tests}
\title{Register the default controller for Syberia project tests.}
\usage{
register_tests(project)
}
\arguments{
\item{project}{director. The syberia director object on which to register
the tests controller.}
}
\description{
By default, tests in a Syberia project will have access to the resource
they are testing as well the \code{testthatsomemore} package.
This function is responsible for bootstrapping this behavior in a
Syberia project.
}

