% Generated by roxygen2 (4.0.1): do not edit by hand
\name{syberia_project}
\alias{syberia_project}
\title{Fetch a syberia project director relative a filename.}
\usage{
syberia_project(filename)
}
\arguments{
\item{filename}{character. Some in a syberia project.}
}
\value{
the director object for the syberia project.
}
\description{
Fetch a syberia project director relative a filename.
}
\examples{
\dontrun{
  # Pretend you have the file blah/foo.R in your syberia project ~/proj.
  syberia_projects('~/proj/blah/foo.R')
  # Now we have the director object for the ~/proj syberia project.
}
}

