% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/tests.r
\name{test_environment_config}
\alias{test_environment_config}
\title{Get the configuration for the test environment.}
\usage{
test_environment_config(project)
}
\arguments{
\item{project}{director or character. The director for the syberia project.}
}
\description{
Get the configuration for the test environment.
}

